requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.5.0
bcrypt==4.2.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
httptools==0.6.1
idna==3.10
importlib_metadata==8.5.0
macholib==1.16.3
packaging==24.1
passlib==1.7.4
pydantic==2.9.2
pydantic_core==2.23.4
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.10
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
zipp==3.20.2
